Rename common to relational_base
